generify listener actors
