+++ src/presentation/cluster_in/mod.rs
@@ -2,21 +2,18 @@ pub mod communication_manager;
 pub mod connection_manager;
 pub mod inbound;
 pub mod outbound;
-pub mod peer_listener;
+
 pub mod peer_listeners;
 
-use std::marker::PhantomData;
-
-use peer_listener::PeerListener;
 use tokio::{net::TcpStream, sync::mpsc::Sender};
 
 use crate::services::cluster::{
     actors::commands::ClusterCommand,
     peers::{
-        connected_types::{FromMaster, FromPeer, FromSlave, ReadConnected, WriteConnected},
+        connected_types::{FromMaster, FromPeer, FromSlave, WriteConnected},
         identifier::PeerIdentifier,
         kind::PeerKind,
-        peer::{ListeningActorKillTrigger, Peer},
+        peer::Peer,
     },
 };
 use crate::services::statefuls::snapshot::snapshot_applier::SnapshotApplier;
@@ -30,36 +27,27 @@ fn create_peer(
 ) -> Peer {
     let (r, w) = stream.into_split();
 
-    let (kill_trigger, kill_switch) = tokio::sync::oneshot::channel();
-
     match kind {
-        PeerKind::Peer => {
-            let rc = ReadConnected::<FromPeer>::new(r);
-            let listening_actor = PeerListener::new(rc, cluster_handler, addr, snapshot_applier);
-            let listener_kill_trigger = ListeningActorKillTrigger::new(
-                kill_trigger,
-                tokio::spawn(listening_actor.listen(kill_switch)),
-            );
-            Peer::new(WriteConnected { stream: w, kind }, listener_kill_trigger)
-        }
-        PeerKind::Replica => {
-            let rc = ReadConnected::<FromSlave>::new(r);
-            let listening_actor = PeerListener::new(rc, cluster_handler, addr, snapshot_applier);
-
-            let listener_kill_trigger = ListeningActorKillTrigger::new(
-                kill_trigger,
-                tokio::spawn(listening_actor.listen(kill_switch)),
-            );
-            Peer::new(WriteConnected { stream: w, kind }, listener_kill_trigger)
-        }
-        PeerKind::Master => {
-            let rc = ReadConnected::<FromMaster>::new(r);
-            let listening_actor = PeerListener::new(rc, cluster_handler, addr, snapshot_applier);
-            let listener_kill_trigger = ListeningActorKillTrigger::new(
-                kill_trigger,
-                tokio::spawn(listening_actor.listen(kill_switch)),
-            );
-            Peer::new(WriteConnected { stream: w, kind }, listener_kill_trigger)
-        }
+        PeerKind::Peer => Peer::new::<FromPeer>(
+            WriteConnected { stream: w, kind },
+            r,
+            cluster_handler,
+            addr,
+            snapshot_applier,
+        ),
+        PeerKind::Replica => Peer::new::<FromSlave>(
+            WriteConnected { stream: w, kind },
+            r,
+            cluster_handler,
+            addr,
+            snapshot_applier,
+        ),
+        PeerKind::Master => Peer::new::<FromMaster>(
+            WriteConnected { stream: w, kind },
+            r,
+            cluster_handler,
+            addr,
+            snapshot_applier,
+        ),
     }
 }
