+++ src/services/cluster/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod actors;
+pub mod peer_listeners;
 pub(crate) mod peers;
 pub(crate) mod replications;
