+++ src/services/cluster/peers/peer.rs
@@ -1,15 +1,19 @@
 use std::ops::Deref;
 use std::ops::DerefMut;
 
-use crate::services::cluster::peers::connected_types::WriteConnected;
+use super::connected_types::ReadConnected;
+use super::identifier::PeerIdentifier;
 
+use crate::services::cluster::actors::commands::ClusterCommand;
+use crate::services::cluster::peer_listeners::PeerListener;
+use crate::services::cluster::peer_listeners::TListen;
+use crate::services::cluster::peers::connected_types::WriteConnected;
+use crate::services::statefuls::snapshot::snapshot_applier::SnapshotApplier;
 use tokio::net::tcp::OwnedReadHalf;
 use tokio::net::tcp::OwnedWriteHalf;
-
+use tokio::sync::mpsc::Sender;
 use tokio::task::JoinHandle;
 use tokio::time::Instant;
-
-use super::connected_types::ReadConnected;
 
 #[derive(Debug)]
 pub(crate) struct Peer {
@@ -19,10 +23,23 @@ pub(crate) struct Peer {
 }
 
 impl Peer {
-    pub fn new(
+    pub fn new<T>(
         write_connected: WriteConnected,
-        listener_kill_trigger: ListeningActorKillTrigger,
-    ) -> Self {
+        read_connected: OwnedReadHalf,
+        cluster_handler: Sender<ClusterCommand>,
+        addr: PeerIdentifier,
+        snapshot_applier: SnapshotApplier,
+    ) -> Self
+    where
+        PeerListener<T>: TListen + Send + Sync + 'static,
+    {
+        let (kill_trigger, kill_switch) = tokio::sync::oneshot::channel();
+        let rc = ReadConnected::<T>::new(read_connected);
+        let listening_actor = PeerListener::new(rc, cluster_handler, addr, snapshot_applier);
+        let listener_kill_trigger = ListeningActorKillTrigger::new(
+            kill_trigger,
+            tokio::spawn(listening_actor.listen(kill_switch)),
+        );
         Self { w_conn: write_connected, listener_kill_trigger, last_seen: Instant::now() }
     }
 }
